+++ packages/expo-file-system/android/src/main/java/expo/modules/filesystem/FileSystemModule.kt
@@ -28,7 +28,6 @@ import kotlinx.coroutines.withContext
 import okhttp3.Call
 import okhttp3.Callback
 import okhttp3.Headers
-import okhttp3.JavaNetCookieJar
 import okhttp3.MediaType
 import okhttp3.MediaType.Companion.toMediaTypeOrNull
 import okhttp3.MultipartBody
@@ -54,7 +53,6 @@ import java.io.InputStream
 import java.io.OutputStream
 import java.io.OutputStreamWriter
 import java.math.BigInteger
-import java.net.CookieHandler
 import java.net.URLConnection
 import java.util.*
 import java.util.concurrent.TimeUnit
@@ -997,10 +995,7 @@ open class FileSystemModule : Module() {
           .connectTimeout(60, TimeUnit.SECONDS)
           .readTimeout(60, TimeUnit.SECONDS)
           .writeTimeout(60, TimeUnit.SECONDS)
-        val cookieHandler: CookieHandler = appContext.legacyModule()
-          ?: throw CookieHandlerNotFoundException()
 
-        builder.cookieJar(JavaNetCookieJar(cookieHandler))
         client = builder.build()
       }
       return client